edit

+++ app/customers/page.jsx
@@ -6,7 +6,8 @@ import Link from 'next/link';
 import { DeleteIcon, HomeIcon } from '@/components/Icons';
 
 function ListaClientes() {
-	const [customers, setCustomers] = useState([]); // Estado para los clientes
+	const [customers, setCustomers] = useState([]); // Estado para los clientes filtrados
+	const [allCustomers, setAllCustomers] = useState([]); // Estado para todos los clientes
 	const [message, setMessage] = useState(null);
 
 	// Función para obtener clientes desde Firebase
@@ -15,14 +16,21 @@ function ListaClientes() {
 			const response = await fetch('/api/getCustomers', { method: 'GET' });
 			if (!response.ok) throw new Error('Error al obtener los clientes');
 			const data = await response.json();
-			setCustomers(data); // Actualiza el estado con los datos obtenidos
+			setCustomers(data); // Actualiza el estado con los datos filtrados
+			setAllCustomers(data); // Almacena todos los clientes en un estado separado
 		} catch (error) {
 			console.error(error.message);
 		}
 	};
 
-	// Función para eliminar un cliente
 	const deleteCustomer = async (id) => {
+		// Mostrar alerta de confirmación
+		const confirmDelete = window.confirm(
+			'¿Estás seguro de que deseas eliminar este cliente?'
+		);
+
+		if (!confirmDelete) return; // Si el usuario cancela, no se ejecuta la eliminación
+
 		try {
 			const response = await fetch(`/api/deleteCustomer`, {
 				method: 'DELETE',
@@ -47,12 +55,48 @@ function ListaClientes() {
 		getCustomers();
 	}, []);
 
+	const handleSearch = (e) => {
+		const searchValue = e.target.value.toLowerCase();
+		if (searchValue === '') {
+			setCustomers(allCustomers); // Si no hay texto de búsqueda, muestra todos los clientes
+		} else {
+			const filteredCustomers = allCustomers.filter((customer) =>
+				customer.cliente
+					.toLowerCase()
+					.slice(0, searchValue.length)
+					.includes(searchValue)
+			);
+			setCustomers(filteredCustomers); // Muestra los clientes filtrados
+		}
+	};
+
 	return (
 		<div className={styles.customersContainer}>
 			<Link href='/home'>
 				<HomeIcon />
 			</Link>
 			<h1 className={styles.title}>Lista de Clientes</h1>
+			<div className={styles.searchContainer}>
+				<input
+					type='text'
+					placeholder='Buscar cliente'
+					className={styles.searchInput}
+					onChange={handleSearch} // Ejecuta la búsqueda al escribir
+				/>
+			</div>
+
+			<Link
+				href='/customers/addCustomer'
+				style={{
+					marginTop: '1rem',
+					marginBottom: '1rem',
+					display: 'block',
+					textAlign: 'center',
+					color: '#1a73e8',
+				}}
+			>
+				Agregar Cliente
+			</Link>
 			{message && <p className={styles.message}>{message}</p>}
 			<div className={styles.cardsContainer}>
 				{customers.map((cliente) => (
@@ -61,6 +105,9 @@ function ListaClientes() {
 						className={styles.customerCard}
 					>
 						<h2 className={styles.customerName}>{cliente.cliente}</h2>
+						<p className={styles.detail}>
+							<strong>Cédula:</strong> {cliente.cedula}
+						</p>
 						<p className={styles.detail}>
 							<strong>Teléfono:</strong> {cliente.teléfono}
 						</p>
@@ -88,17 +135,6 @@ function ListaClientes() {
 					</div>
 				))}
 			</div>
-			<Link
-				href='/customers/addCustomer'
-				style={{
-					marginTop: '1rem',
-					display: 'block',
-					textAlign: 'center',
-					color: '#1a73e8',
-				}}
-			>
-				Agregar Cliente
-			</Link>
 		</div>
 	);
 }
